requirements.txt
aiogram>=3.0.0,<4.0.0
google-generativeai>=0.3.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
